Build time variable
I may have screwed things in my local so I am pushing the only changes I know are fine, so this is a push pull kind of moment. The build time variable has been implemented into the data asset now and all building have an exposed build time variable with a function being created in the BP_PlacementManager to check if it has surpassed the build time to start the resource gatehring.