Building delay is so much fun
Currently I have added temp values to the data assets and gave them the corresponding data assets in the blueprints to access the values. The values are able to be stored within the struct function, I am currently working on trying to update the values after placement at the quarters and only on the quarters.