Needs Testing
Have it implemented such that a skill tree node isn't selectable until you have the right soul cost, and then you have to pay it when you select it. I need a way to display that the cost isn't met though, or a way to view how many souls the player has. The first skill is greyed out, even when the other prerequisites are met, but I can't exactly confirm why just yet.