Added art scene menu button
There's a button to access the art scene from the main menu now. It seems to work when supplied with the name of another scene, but I'm assuming the name of the art scene is VS_ArtScene since it's not on the main branch yet. If it's not, the function will need to be updated.