Fixed most of the settings feedback
I was able to clear up most of the issues with the settings widget. I'm gonna have to ask about getting a master audio bus before I can fully optimize the audio settings, and I'm still not sure how to fix the audio output select device, but I took care of everything else. I might, in the future, want to make sure the tab the widget starts on has it's button disabled immediately, but that has no functional consequence, so I might save it for when I improve the audio settings again. Sorry this is so late, also.