Building widgets implemented
The buildings progress bar widget has been implemented and is now functional.

As time goes by the widget fills up as a progress bar which pauses when we pause the game or have a pause of any type in the game.

In regards to what has been changed in the structure base

I have added the widget which displays the progress bar, it updates from the WBP_Structure_Progress. I added the data initialization for the progress bar widget inside of the StructureBase initialization at the end. I threw a comment over it to show where it is initialized.

Inside of the WBP_Structure_Progress I added in the Functionality of the percentage bar which is inside the get percentage function essentially checks how much time has progress compared to the amount of time it takes for the building to be completed.

The event tick while it may be scary for larger amounts of buildings disables itself when the building progress is done.

This lower function in the event graph checks to see if the game is paused and if the building is done building. Consolidated the info into macros to help with visibility or else it was gonna start looking spaghettish and we don't like that.

In each of the Building blueprints I took the liberty of centering the widget above the buildings placement with the exception of the citadel where I disabled the visibility altogether because we don't need it to show any build time.