Updated Building progress bar widget
I have gone ahead and fixed up the pause feature which now checks the game menu as the events use up this menu if I am not mistaken. it also checks if we the player paused it which would override the game menu by making it false.

The data asset was set improperly in the getData asset in path which also caused it to get a little wonky. Should be good now.

This should have taken care of both the bugs that we had going on with this originally.