Scaffolding, and adjustments to the looks
Has all the assets that are being added into the game and some adjustments to the effects making them centered.

I did also move the files for the older Structures into a folder called old structures just to make it known we don't use those.

Only things that were changed coding wise was inside the bp_structure_base which is now enabling the static mesh for the scaffolding to be visible and then it becomes invisible when its done.

I did have to adjust the size of the models I may have done an unconventional approach to it and made it so the models them selves were reduced.

The only other addition was adding the static mesh to the primary data asset which I have update the assets to have their respective meshes.