Fixed bugs according to:
Can you explain why leeway and difficulty are in one manager and peril is in its own? In my head leeway and peril are the things that trigger loss conditions and the ones that should be in the same manager.  Not saying you have to change it I'm just curious the reason
So peril goes up if the player has 0 people OR order is below the threshold? wanna run this by @Harrison Weber .
Is the call to start peril at 5 seconds and stop it at 20 seconds just for demonstration purposes?
The soul quota stays at 0 for the whole game. It needs a default value of 10.
When I lost the game the menu got triggered repetedly and the sound was spammed. Make sure the final "lose game" function only gets called once when peril is up.