Initial Commit for improving build communication
This removes the grid when the building is placed however it does not do it immediately however after we place the building, the next time we open the build menu it won't show the grid tiles under a building.

Also added in the materials for being able to place the building not fully implemented as well, need to tweak it so the building scaffolding is more transparent when we are trying to place it.